+++ src/com/rakovets/course/challenge/objects/Task04_Rectangle.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.challenge.oop;
+package com.rakovets.course.challenge.objects;
 
 /*## Specification of task 4
 Создать класс `Rectangle`.
